# perf(validation): modernize commit validation with Bun runtime and industry standards